Added serialization/deserialization unit tests for request objects.

+++ OpenAISharp.Edit/Requests/CreateEditRequest.cs
@@ -56,7 +56,7 @@ namespace OpenAISharp.Edit.Requests
         /// <remarks>https://beta.openai.com/docs/api-reference/edits/create#edits/create-temperature</remarks>
         [JsonPropertyName("temperature")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public int? Temperature { get; set; }
+        public double? Temperature { get; set; }
 
         /// <summary>
         /// An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. 
@@ -66,7 +66,7 @@ namespace OpenAISharp.Edit.Requests
         /// <remarks>https://beta.openai.com/docs/api-reference/edits/create#edits/create-top_p</remarks>
         [JsonPropertyName("top_p")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public int? TopP { get; set; }
+        public double? TopP { get; set; }
     }
 }
